+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [vNext]
 ### Added
+- Implemented test on getting status bar height. [Trello 1478](https://trello.com/c/RuPL3v4v)
 - Implemented test on setting deviceName to appEnvironment data. [Trello 125](https://trello.com/c/ekZqajRU)
 - Supporting `VisualGridOptions` in configuration and in fluent API. [Trello 2089](https://trello.com/c/d4zggQes)
 ### Updated